move getuserlocation to helpers
